requirements.txt
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
gradio==5.31.0
gradio_client==1.10.1
groovy==0.1.2
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.41.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
plotly==6.1.1
pluggy==1.6.0
protobuf==6.31.0
psutil==7.0.0
pvlib==0.12.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
 pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.1
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
ruff==0.11.11
safehttpx==0.1.6
scipy==1.15.3
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.2
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.145
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
# open3d==0.18.0  # Not available for Python 3.13 yet
scikit-learn==1.3.2
